Move generic url functions to helpers
