+++ projects/Mallard/src/authentication/helpers.ts
@@ -10,20 +10,6 @@ import {
     TokenType,
 } from 'src/services/id-service'
 
-const parseSearchString = (searchString: string) => {
-    const paramString = searchString.split('&')
-    const init: { [key: string]: string } = {}
-    return paramString.reduce((acc, p) => {
-        const [key, value] = p.split('=')
-        return { ...acc, [key]: decodeURIComponent(value) }
-    }, init)
-}
-
-const createSearchParams = (params: { [key: string]: string } = {}) =>
-    Object.entries(params)
-        .map(([key, value]) => `${key}=${value}`)
-        .join('&')
-
 const fetchAndPersistUserAccessTokenWithIdentity = async (
     email: string,
     password: string,
@@ -64,6 +50,4 @@ export {
     fetchAndPersistUserAccessTokenWithIdentity,
     fetchAndPersistUserAccessTokenWithType,
     fetchMembershipDataForKeychainUser,
-    parseSearchString,
-    createSearchParams,
 }
